rename: rooms 에 대한 파일 이름을 roomList로 변경
