+++ src/domain/rooms/components/RoomList.tsx
@@ -11,8 +11,8 @@ import { ActivityIndicator } from "react-native-paper";
 import { RFValue } from "react-native-responsive-fontsize";
 import { useSelectedRoomStore } from "../../prayerRoom/types/roomStore";
 import { Room } from "../types/dto/responses/room";
-import { useToggleRoomNotificationMutation } from "./../hooks/mutations/roomMutations";
-import { useInfiniteRoomsQuery } from "./../hooks/queries/roomQueries";
+import { useToggleRoomNotificationMutation } from "../hooks/mutations/roomListMutations";
+import { useInfiniteRoomsQuery } from "../hooks/queries/roomListQueries";
 import EmptyRoomList from "./RoomEmpty";
 import RoomItem from "./RoomItem";
 import { useRouter } from "expo-router";
